backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
langchain-community==0.3.7
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
pypdf==4.0.2

